requirements.txt
pandas
openpyxl
opencv-python
matplotlib
PyQt6
pillow
torchvision
tqdm
wandb
tensorboard
seaborn
scikit-learn
torchmetrics
lightning